move difficulites to constants.js
