+++ src/App.js
@@ -2,24 +2,6 @@ import React, { Component } from "react";
 import Landing from "./pages/Landing";
 import Game from "./pages/Game";
 
-const difficulties = {
-  easy: {
-    key: "easy",
-    label: "Easy",
-    factor: 1,
-  },
-  medium: {
-    key: "medium",
-    label: "Medium",
-    factor: 1.5,
-  },
-  hard: {
-    key: "hard",
-    label: "Hard",
-    factor: 2,
-  },
-};
-
 class App extends Component {
   state = {
     gameStarted: false,
@@ -45,10 +27,9 @@ class App extends Component {
             user={{ name: this.state.name }}
             difficulty={this.state.difficulty}
             onEnd={this.handleGameEnd}
-            difficulties={difficulties}
           />
         ) : (
-          <Landing onStart={this.handleGameStart} difficulties={difficulties} />
+          <Landing onStart={this.handleGameStart} />
         )}
       </main>
     );
