+++ src/constants.js
@@ -1,0 +1,19 @@
+const difficulties = {
+  easy: {
+    key: "easy",
+    label: "Easy",
+    factor: 1,
+  },
+  medium: {
+    key: "medium",
+    label: "Medium",
+    factor: 1.5,
+  },
+  hard: {
+    key: "hard",
+    label: "Hard",
+    factor: 2,
+  },
+};
+
+export { difficulties };
